REFACTOR LogsTable to own template
