+++ pwa/src/templates/logsTemplate/LogsTemplate.module.css
@@ -2,61 +2,6 @@
   margin-block-end: var(--gateway-ui-size-2xl);
 }
 
-.tableRow {
-  cursor: pointer;
-}
-
-.button > span > svg {
-  margin-inline-end: var(--gateway-ui-size-xs);
-}
-
-.message {
-  max-width: 200px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.message > :first-child {
-  width: 100%;
-}
-
-.DEBUG {
-  background-color: var(--gateway-ui-color-log-level-debug);
-  color: darken(var(--gateway-ui-color-log-level-debug), 50%);
-}
-
-.INFO {
-  background-color: var(--gateway-ui-color-log-level-info);
-}
-
-.NOTICE {
-  background-color: var(--gateway-ui-color-log-level-notice);
-}
-
-.WARNING {
-  background-color: var(--gateway-ui-color-log-level-warning);
-}
-
-.ERROR {
-  background-color: var(--gateway-ui-color-log-level-error);
-}
-
-.CRITICAL {
-  background-color: var(--gateway-ui-color-log-level-critical);
-  color: var(--gateway-ui-color-white);
-}
-
-.ALERT {
-  background-color: var(--gateway-ui-color-log-level-alert);
-  color: var(--gateway-ui-color-white);
-}
-
-.EMERGENCY {
-  background-color: var(--gateway-ui-color-log-level-emergency);
-  color: var(--gateway-ui-color-white);
-}
-
 /* pagination */
 
 .paginationContainer {
